Fixed code duplication
